+++ src/Curiosity.Email.UnisenderGo/IUnisenderGoEmailSender.cs
@@ -1,0 +1,11 @@
+using Curiosity.EMail;
+
+namespace Curiosity.Email.UnisenderGo
+{
+    /// <summary>
+    /// Class for sending Emails via UnisenderGo.
+    /// </summary>
+    public interface IUnisenderGoEmailSender : IEMailSender
+    {
+    }
+}