+++ src/Curiosity.Email.UnisenderGo/UnisenderGoRegion.cs
@@ -1,0 +1,13 @@
+namespace Curiosity.Email.UnisenderGo
+{
+    /// <summary>
+    /// Region where UnisenderGo API is located.
+    /// </summary>
+    public enum UnisenderGoRegion
+    {
+        /// <summary>
+        /// Russian Federation.
+        /// </summary>
+        Russia = 0
+    }
+}